Fixed Typos.
Fixed a sentence that used a word that did not fit. I think the author
means to say that these principles can not be altered/customized by a
team, but used a reverse logic. I hope this change solidifies the
relationship between the team and the principles in a way that doesn't
seem imutable. The principles can chage over time, but through mindful
agreement of the teams that uphold this manifesto.

The other change was a spelling correction.